% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_plot.R
\name{contour_plot}
\alias{contour_plot}
\title{Plot the contour grid}
\usage{
contour_plot(grid_shape, background_col = "#2a3c4b", line_col = "#e9ebed")
}
\arguments{
\item{grid_shape}{a dataframe of x, y, and z points}

\item{background_col}{a character string used for the background color of the plot}

\item{line_col}{a character string used for the line colors of your plot}
}
\value{
a ggplot2 object with some default parameters of line size of 0.25,
linejoin is round, and uses theme_void() to remove extra elements
}
\description{
Plot the contour grid
}
