% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_shape.R
\name{contour_shape}
\alias{contour_shape}
\title{Turn the contour grid into a shape}
\usage{
contour_shape(grid, radius, x_center, y_center)
}
\arguments{
\item{grid}{dataframe of x, y, and z points}

\item{radius}{radius of the shape}

\item{x_center}{x coordinate of the center of the shape}

\item{y_center}{y coordinate of the center of the shape}
}
\value{
data frame with the x, y and z coordinates of the points
}
\description{
Turn the contour grid into a shape
}
